+++ include/libcyphal/runnable.hpp
@@ -1,6 +1,3 @@
-/// @file
-/// libcyphal common header.
-///
 /// @copyright
 /// Copyright (C) OpenCyphal Development Team  <opencyphal.org>
 /// Copyright Amazon.com Inc. or its affiliates.
@@ -14,24 +11,14 @@
 namespace libcyphal
 {
 
-/// @brief Declares an abstract Cyphal runnable interface.
-///
-/// Runnable objects do work asynchronously but only when `run()` is called.
-/// This allows super-loop firmware to separate execution into application work and library work,
-/// and for threaded software applications to service the library from dedicated threads.
-/// Each object that implements `IRunnable` shall document how when it must be run to achieve certain functionality and
-/// timing guarantees.
-///
 class IRunnable
 {
 public:
-    /// @brief Runs the runnable object.
     virtual void run(TimePoint now) = 0;
 
 protected:
     virtual ~IRunnable() = default;
-
-};  // IRunnable
+};
 
 }  // namespace libcyphal
