+++ include/libcyphal/transport/session/msg_sessions.hpp
@@ -1,6 +1,3 @@
-/// @file
-/// Defines the Message Session interfaces for Transport Layer implementations.
-///
 /// @copyright
 /// Copyright (C) OpenCyphal Development Team  <opencyphal.org>
 /// Copyright Amazon.com Inc. or its affiliates.
@@ -15,44 +12,31 @@ namespace libcyphal
 {
 namespace transport
 {
-inline namespace session
+namespace session
 {
 
-/// @brief Declares message RX session parameters.
-///
 struct MessageRxParams final
 {
-    PortId      subject_id;
     std::size_t extent_bytes;
+    PortId      subject_id;
+};
 
-};  // MessageRxParams
-
-/// @brief Declares message TX session parameters.
-///
 struct MessageTxParams final
 {
     PortId subject_id;
+};
 
-};  // MessageTxParams
-
-/// @brief Declares an abstract Cyphal transport message RX session interface.
-///
-class IMessageRxSession : public IRunnable
+class IMessageRxSession : public IRxSession
 {
 public:
-    CETL_NODISCARD virtual MessageRxParams getParams() const noexcept             = 0;
-    virtual void                           setTransferIdTimeout(Duration timeout) = 0;
+    CETL_NODISCARD virtual MessageRxParams getParams() const noexcept = 0;
+};
 
-};  // IMessageRxSession
-
-/// @brief Declares an abstract Cyphal transport message TX session interface.
-///
 class IMessageTxSession : public IRunnable
 {
 public:
     CETL_NODISCARD virtual MessageTxParams getParams() const noexcept = 0;
-
-};  // IMessageTxSession
+};
 
 }  // namespace session
 }  // namespace transport
