+++ include/libcyphal/transport/session/svc_sessions.hpp
@@ -1,6 +1,3 @@
-/// @file
-/// Defines the Service Session interfaces for Transport Layer implementations.
-///
 /// @copyright
 /// Copyright (C) OpenCyphal Development Team  <opencyphal.org>
 /// Copyright Amazon.com Inc. or its affiliates.
@@ -15,82 +12,56 @@ namespace libcyphal
 {
 namespace transport
 {
-inline namespace session
+namespace session
 {
 
-/// @brief Declares service request RX session parameters.
-///
 struct RequestRxParams final
 {
-    PortId      service_id;
     std::size_t extent_bytes;
+    PortId      service_id;
+};
 
-};  // RequestRxParams
-
-/// @brief Declares service request TX session parameters.
-///
 struct RequestTxParams final
 {
     PortId service_id;
     NodeId server_node_id;
+};
 
-};  // RequestTxParams
-
-/// @brief Declares service response RX session parameters.
-///
 struct ResponseRxParams final
 {
+    std::size_t extent_bytes;
     PortId      service_id;
     NodeId      server_node_id;
-    std::size_t extent_bytes;
+};
 
-};  // ResponseRxParams
-
-/// @brief Declares service response TX session parameters.
-///
 struct ResponseTxParams final
 {
     PortId service_id;
+};
 
-};  // ResponseTxParams
-
-/// @brief Declares an abstract Cyphal transport request RX session interface.
-///
-class IRequestRxSession : public ISession
+class IRequestRxSession : public IRxSession
 {
 public:
-    CETL_NODISCARD virtual RequestRxParams getParams() const noexcept             = 0;
-    virtual void                           setTransferIdTimeout(Duration timeout) = 0;
+    CETL_NODISCARD virtual RequestRxParams getParams() const noexcept = 0;
+};
 
-};  // IRequestRxSession
-
-/// @brief Declares an abstract Cyphal transport request TX session interface.
-///
 class IRequestTxSession : public ISession
 {
 public:
     CETL_NODISCARD virtual RequestTxParams getParams() const noexcept = 0;
+};
 
-};  // IRequestTxSession
-
-/// @brief Declares an abstract Cyphal transport response RX session interface.
-///
-class IResponseRxSession : public ISession
+class IResponseRxSession : public IRxSession
 {
 public:
-    CETL_NODISCARD virtual ResponseRxParams getParams() const noexcept             = 0;
-    virtual void                            setTransferIdTimeout(Duration timeout) = 0;
+    CETL_NODISCARD virtual ResponseRxParams getParams() const noexcept = 0;
+};
 
-};  // IResponseRxSession
-
-/// @brief Declares an abstract Cyphal transport response TX session interface.
-///
 class IResponseTxSession : public ISession
 {
 public:
     CETL_NODISCARD virtual ResponseTxParams getParams() const noexcept = 0;
-
-};  // IResponseTxSession
+};
 
 }  // namespace session
 }  // namespace transport
